Reduce frequency of failures due to transient files in the shared dir
Unfortunately, this will also have occasional issues because chown -R will
still fail if a file is removed during the recursive change of ownership.

However, because it is faster than the non-recursive chown, it should
reduce the frequency of errors.

chown --silent would avoid the problem, but what other errors are then
ignored siliently?

Find examples:
  Example directory structure:
    $ find -print
      .
      ./cached-copy
      ./cached-copy/cached_file
      ./bundle
      ./bundle/bundle_file
      ./some_dir
      ./some_dir/other_file.txt
      ./some_file.txt

  Output of current find command:
    $ find . -name cached-copy -prune -o -name bundle -prune -o -print
      .
      ./some_dir
      ./some_dir/other_file.txt
      ./some_file.txt

  Output of updated find command:
    find . -maxdepth 1 -not -name bundle -not -name cached-copy -print
      .
      ./some_dir
      ./some_file.txt